fix: separate client and server Supabase instances
